Scheduler added, check if empty fields
